Add files via upload
adult_sample_with_header.csv is the file created by strictly following the steps in datacamp. However, this file has only the header, which I believe is caused by a bug in the code.
adult_sample_with_header_revised.csv is the file created based on a few changes to the code in datacamp according to my own understanding of what that link wants us to do.